Externalizing TableHeader
