+++ kie-wb-common-react/boxed-expression-editor/src/components/Table/Table.tsx
@@ -21,14 +21,12 @@ import {
   ColumnInstance,
   ContextMenuEvent,
   DataRecord,
-  HeaderGroup,
   Row,
   useBlockLayout,
   useResizeColumns,
   useTable,
 } from "react-table";
-import { TableComposable, Tbody, Td, Th, Thead, Tr } from "@patternfly/react-table";
-import { EditExpressionMenu } from "../EditExpressionMenu";
+import { TableComposable, Tbody, Td, Tr } from "@patternfly/react-table";
 import * as React from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { EditableCell } from "./EditableCell";
@@ -36,6 +34,7 @@ import { CellProps, DataType, TableHandlerConfiguration, TableOperation } from "
 import * as _ from "lodash";
 import { Popover } from "@patternfly/react-core";
 import { TableHandlerMenu } from "./TableHandlerMenu";
+import { TableHeader } from "./TableHeader";
 
 export interface TableProps {
   /** Table identifier, useful for nested structures */
@@ -138,38 +137,6 @@ export const Table: React.FunctionComponent<TableProps> = ({
   const deleteAt = <T extends unknown>(elements: T[], index: number) => {
     return [...elements.slice(0, index), ...elements.slice(index + 1)];
   };
-
-  const updateColumnNameInRows = useCallback(
-    (prevColumnName: string, newColumnName: string) =>
-      setTableRows((prevTableCells) => {
-        return _.map(prevTableCells, (tableCells) => {
-          const assignedCellValue = tableCells[prevColumnName]!;
-          delete tableCells[prevColumnName];
-          tableCells[newColumnName] = assignedCellValue;
-          return tableCells;
-        });
-      }),
-    []
-  );
-
-  const onColumnNameOrDataTypeUpdate = useCallback(
-    (columnIndex: number) => {
-      return ({ name = "", dataType = DataType.Undefined }) => {
-        const prevColumnName = (tableColumns[columnIndex] as ColumnInstance).accessor as string;
-        setTableColumns((prevTableColumns: ColumnInstance[]) => {
-          const updatedTableColumns = [...prevTableColumns];
-          updatedTableColumns[columnIndex].label = name;
-          updatedTableColumns[columnIndex].accessor = name;
-          updatedTableColumns[columnIndex].dataType = dataType;
-          return updatedTableColumns;
-        });
-        if (name !== prevColumnName) {
-          updateColumnNameInRows(prevColumnName, name);
-        }
-      };
-    },
-    [tableColumns, updateColumnNameInRows]
-  );
 
   const onCellUpdate = useCallback((rowIndex: number, columnId: string, value: string) => {
     setTableRows((prevTableCells) => {
@@ -373,68 +340,6 @@ export const Table: React.FunctionComponent<TableProps> = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [finishedResizing]);
 
-  const renderCountColumn = useCallback(
-    (column: ColumnInstance, columnIndex: number) => (
-      <Th {...column.getHeaderProps()} className="fixed-column no-clickable-cell" key={columnIndex}>
-        <div className="header-cell" data-ouia-component-type="expression-column-header">
-          {column.label}
-        </div>
-      </Th>
-    ),
-    []
-  );
-
-  const renderResizableHeaderCell = useCallback(
-    (column, columnIndex) => (
-      <Th
-        {...column.getHeaderProps()}
-        {...tableInstance.getThProps(column, columnIndex)}
-        className={`resizable-column ${!column.dataType ? "no-clickable-cell" : null}`}
-        key={columnIndex}
-      >
-        <div className="header-cell" data-ouia-component-type="expression-column-header">
-          <div>
-            <p className="pf-u-text-truncate">{column.label}</p>
-            {column.dataType ? <p className="pf-u-text-truncate data-type">({column.dataType})</p> : null}
-          </div>
-          <div
-            className={`pf-c-drawer ${!column.canResize ? "resizer-disabled" : ""}`}
-            {...(column.canResize ? column.getResizerProps() : {})}
-          >
-            <div className="pf-c-drawer__splitter pf-m-vertical">
-              <div className="pf-c-drawer__splitter-handle" />
-            </div>
-          </div>
-        </div>
-      </Th>
-    ),
-    [tableInstance]
-  );
-
-  const renderResizableColumn = useCallback(
-    (column: ColumnInstance, columnIndex: number) =>
-      column.dataType ? (
-        <EditExpressionMenu
-          title={editColumnLabel}
-          selectedExpressionName={column.label}
-          selectedDataType={column.dataType}
-          onExpressionUpdate={onColumnNameOrDataTypeUpdate(columnIndex)}
-          key={columnIndex}
-        >
-          {renderResizableHeaderCell(column, columnIndex)}
-        </EditExpressionMenu>
-      ) : (
-        renderResizableHeaderCell(column, columnIndex)
-      ),
-    [editColumnLabel, onColumnNameOrDataTypeUpdate, renderResizableHeaderCell]
-  );
-
-  const renderColumn = useCallback(
-    (column: ColumnInstance, columnIndex: number) =>
-      column.isCountColumn ? renderCountColumn(column, columnIndex) : renderResizableColumn(column, columnIndex),
-    [renderCountColumn, renderResizableColumn]
-  );
-
   const renderAdditiveRow = useMemo(
     () => (
       <Tr className="table-row additive-row">
@@ -461,32 +366,17 @@ export const Table: React.FunctionComponent<TableProps> = ({
     [children, tableInstance.allColumns]
   );
 
-  const renderHeaderGroups = useMemo(
-    () =>
-      tableInstance.headerGroups.map((headerGroup: HeaderGroup) => (
-        <Tr key={headerGroup.accessor} {...headerGroup.getHeaderGroupProps()}>
-          {headerGroup.headers.map((column: ColumnInstance, columnIndex: number) => renderColumn(column, columnIndex))}
-        </Tr>
-      )),
-    [renderColumn, tableInstance.headerGroups]
-  );
-
-  const renderLastLevelInHeaderGroups = useMemo(
-    () => (
-      <Tr>
-        {_.last(
-          tableInstance.headerGroups as HeaderGroup[]
-        )!.headers.map((column: ColumnInstance, columnIndex: number) => renderColumn(column, columnIndex))}
-      </Tr>
-    ),
-    [renderColumn, tableInstance.headerGroups]
-  );
-
   return (
     <div className={`table-component ${tableId}`}>
       <TableComposable variant="compact" {...tableInstance.getTableProps()} ref={tableRef}>
-        <Thead noWrap>{isHeadless ? renderLastLevelInHeaderGroups : renderHeaderGroups}</Thead>
-
+        <TableHeader
+          tableInstance={tableInstance}
+          editColumnLabel={editColumnLabel}
+          isHeadless={isHeadless}
+          tableColumns={tableColumns as ColumnInstance[]}
+          setTableColumns={setTableColumns}
+          setTableRows={setTableRows}
+        />
         <Tbody {...tableInstance.getTableBodyProps()}>
           {tableInstance.rows.map((row: Row, rowIndex: number) => {
             tableInstance.prepareRow(row);
