+++ kie-wb-common-react/boxed-expression-editor/src/components/Table/index.ts
@@ -17,3 +17,4 @@
 export * from "./EditableCell";
 export * from "./Table";
 export * from "./TableHandlerMenu";
+export * from "./TableHeader";
